requirements.txt
pyrogram==2.0.106
pymongo==4.3.3
pytz==2022.1
python-dotenv==0.19.0
pyromod==1.5
